+++ src/Integration/Monolog/CloudWatch/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## NOT RELEASED
+
+### Changed
+
+- BC: `CloudWatchLogsHandler` constructor arguments changed
+
 ## 0.1.0
 
 First version